% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{La fonction sert à afficher Hello, world!
Cela permet de vérifier que tout marche bien}
\usage{
hello()
}
\value{
"Hello, world!"
}
\description{
La fonction sert à afficher Hello, world!
Cela permet de vérifier que tout marche bien
}
\examples{
hello()
}
