% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc.R
\name{smc}
\alias{smc}
\title{renvoie une partie des donnees de smc}
\usage{
smc(df)
}
\value{
renvoie les données issues de la colonne Age et de la ligne 33
}
\description{
renvoie une partie des donnees de smc
}
\examples{
smc()
}
